feat: Add function of wiki edit
