+++ src/components/WIKI/WikiEditor.tsx
@@ -1,70 +1,51 @@
-import React, { useState, useCallback } from "react";
-import { useNavigate, useParams } from "react-router-dom";
-import { Form, Input, Button, message } from "antd";
-import { useWiki } from "../../hooks/useWiki";
+import React, { useState } from 'react';
 
-const { TextArea } = Input;
+const WikiEditor = ({ initialContent, onSave }) => {
+  const [content, setContent] = useState(initialContent.content);
 
-interface WikiFormData {
-  title: string;
-  content: string;
-}
+  // 텍스트 변경 시 상태 업데이트
+  const handleChange = (e) => {
+    setContent(e.target.value);
+  };
 
-export const WikiEditor: React.FC = () => {
-  const { wiki_title } = useParams<{ wiki_title: string }>();
-  const navigate = useNavigate();
-  const { wikiData, loading, updateWiki, createWiki } = useWiki(wiki_title);
-  const [form] = Form.useForm<WikiFormData>();
-
-  const handleSubmit = useCallback(
-    async (values: WikiFormData) => {
-      try {
-        if (wiki_title) {
-          await updateWiki(values);
-          message.success("문서가 수정되었습니다.");
-        } else {
-          await createWiki(values);
-          message.success("문서가 생성되었습니다.");
-        }
-        navigate(`/wiki/${values.title}`);
-      } catch (error) {
-        message.error("문서 저장 중 오류가 발생했습니다.");
-      }
-    },
-    [wiki_title, updateWiki, createWiki, navigate]
-  );
-
-  if (loading) {
-    return <div>Loading...</div>;
-  }
+  // 저장 버튼 클릭 시 호출
+  const handleSave = () => {
+    onSave(content);
+  };
 
   return (
-    <Form
-      form={form}
-      layout="vertical"
-      onFinish={handleSubmit}
-      initialValues={wikiData}
-    >
-      <Form.Item
-        name="title"
-        label="제목"
-        rules={[{ required: true, message: "제목을 입력해주세요" }]}
-      >
-        <Input />
-      </Form.Item>
-      <Form.Item
-        name="content"
-        label="내용"
-        rules={[{ required: true, message: "내용을 입력해주세요" }]}
-      >
-        <TextArea rows={20} />
-      </Form.Item>
-      <Form.Item>
-        <Button type="primary" htmlType="submit">
-          저장
-        </Button>
-      </Form.Item>
-    </Form>
+    <div className="max-w-4xl p-6 mx-auto bg-white rounded-md shadow-md">
+      {/* 에디터 영역 */}
+      <div className="mb-6">
+        <h2 className="mb-4 text-2xl font-semibold text-gray-700">위키 수정</h2>
+        <textarea
+          className="w-full h-96 p-4 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+          value={content}
+          onChange={handleChange}
+        />
+      </div>
+
+      <div className="flex justify-end space-x-4">
+        {/* 도움말 링크 */}
+        <a
+          href="/wiki/도움말" // 도움말 링크
+          className="px-4 py-2 text-white bg-gray-600 rounded-md shadow-md hover:underline hover:bg-gray-700"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          도움말
+        </a>
+        {/* 저장 버튼 */}
+        <button
+          className="px-4 py-2 text-white bg-gray-600 rounded-md shadow-md hover:underline hover:bg-gray-700"
+          onClick={handleSave}
+        >
+          수정
+        </button>
+
+
+      </div>
+    </div>
   );
 };
