+++ src/components/WIKI/WikiEngine.jsx
@@ -3,6 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import ReactDiffViewer from 'react-diff-viewer-continued';
 import { toRelativeTime } from '../../utils/Time';
 import useWindowDimensions from '../../hooks/useWindowDimensions';
+import { Link } from 'react-router-dom';
+import { useAuth } from '../../hooks/useAuth';
 
 const WikiContent = ({ author, DocTitle, content, notFoundFlag, history, prevContent }) => {
   const [toc, setToc] = useState([]);
@@ -11,7 +13,7 @@ const WikiContent = ({ author, DocTitle, content, notFoundFlag, history, prevCon
   const [isContentVisible, setIsContentVisible] = useState(history === undefined ? true : false);
   const [isHistoryVisible, setIsHistoryVisible] = useState(history === undefined ? true : false);
   const navigate = useNavigate();
-  let accessToken = localStorage.getItem("accessToken");
+  const { isLoggedIn } = useAuth();
   const { height, width } = useWindowDimensions();
 
   const applyFormatting = (text) => {
@@ -148,14 +150,14 @@ const WikiContent = ({ author, DocTitle, content, notFoundFlag, history, prevCon
 
     <div className="flex space-x-4">
       <>
-        {accessToken &&
-          <a href={`/wiki/edit/${DocTitle}`} className='px-4 py-2 text-white bg-gray-600 rounded-md shadow-md hover:bg-gray-700'>
+        {isLoggedIn &&
+          <Link to={`/wiki/edit/${DocTitle}`} className='px-4 py-2 text-white bg-gray-600 rounded-md shadow-md hover:bg-gray-700'>
             수정
-          </a>
+          </Link>
         }
-        <a href={`/wiki/history/${DocTitle}`} className='px-4 py-2 text-white bg-gray-600 rounded-md shadow-md hover:bg-gray-700'>
+        <Link to={`/wiki/history/${DocTitle}`} className='px-4 py-2 text-white bg-gray-600 rounded-md shadow-md hover:bg-gray-700'>
           수정 내역
-        </a>
+        </Link>
       </>
     </div>
   );
